+++ crates/oxc_formatter_core/src/lib.rs
@@ -40,7 +40,7 @@ pub mod test_support;
 pub use arguments::{Argument, Arguments};
 pub use buffer::{
     AccumulatorBuffer, Buffer, BufferExtensions, HeapVecBuffer, Inspect, PreambleBuffer, Recorded,
-    Recording, RemoveSoftLinesBuffer, VecBuffer,
+    Recording, RemoveSoftLinesBuffer, ScratchBuffer, VecBuffer,
 };
 pub use diagnostics::{ActualStart, FormatError, InvalidDocumentError, PrintError};
 pub use embedded::{
@@ -67,7 +67,7 @@ pub use options::{
 pub use printer::{PrintResult, PrintWidth, Printed, Printer, PrinterOptions};
 pub(crate) use simple_context::SimpleFormatContext;
 pub use source_text::SourceText;
-pub use state::{FormatState, ScratchBuffer};
+pub use state::FormatState;
 pub use text_range::TextRange;
 pub use traits::{FormatContext, FormatOptions};
